cafesys/brassbird/requirements.txt
django-piston
